requirements.txt
Cython
future
matplotlib
numpy
opencv-python
Pillow
pprint
protobuf
PyYAML